requirements.txt
#boto3==1.35.8
#botocore==1.35.8
#certifi==2024.7.4
#cffi==1.17.0
#charset-normalizer==3.3.2
#cryptography==43.0.0
#idna==3.8
#jmespath==1.0.1
#numpy==2.1.0
#packaging==24.1
#pdf2image==1.17.0
#pdfminer.six==20231228
#pdfplumber==0.11.4
#pillow==10.4.0
#pycparser==2.22
#pypdfium2==4.30.0
#pytesseract==0.3.13
#python-dateutil==2.9.0.post0
#requests==2.32.3
#s3transfer==0.10.2
#six==1.16.0
#urllib3==2.2.2
boto3
langchain
langchain-aws
streamlit
chardet
python-dotenv
langchain-community
